Beta3 -> Beta 4: Make sure album list completely reloads so there is no false file date information
We added a field, creationTimeMS, to ALbum.java. This is used to populate the
"Added" column in the album library pane, and also when selecting "Sort by
Recently Added" on the right click context menu. This makes sure that field
is always accurately populated, by invalidating all previously cache'd-to-disk
album data, forcing a complete rescan for the beta4 when transitioning
from beta3.